[TheNextWeek] Refactor hittable API. - Use the same approach as for textures. - Keep some header files (like Rectangle and Instance) for convenience. - Keep BVH header file to prevent recursive dependency (BVH -> hittable_list -> hittable).

+++ rt_perlin.c
@@ -112,7 +112,7 @@ static int *generate_permutation(size_t size)
     int *permutation = malloc(sizeof(int) * size);
     assert(NULL != permutation);
 
-    for (size_t i = 0; i < size; ++i)
+    for (int i = 0; i < size; ++i)
     {
         permutation[i] = i;
     }
